Create Readme.md file
This app is being developed as a proposed business solution for customers, wineries, and distributors at all three levels, to offer solutions that presently do not exist in the industry.  As someone who has formerly owned a winery, and someone who has worked with distributors regularly for roughly a decade, I have extensive knowledge in this area.  I also have the front facing forward knowledge of working with customers and hearing the voicing of complaints and criticisms of the trade.  While I took each with a grain of salt, I did not forget about those complaints, instead choosing to direct those into a focused model that might one day be used as a revenue generation force in the market.